third_party/WebKit/LayoutTests/http/tests/serviceworker/controller-on-load.html

<!DOCTYPE html>
<title>Service Worker: Controller on load</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<body>
<script>
var t = async_test('controller is set for a controlled document');
t.step(function() {
    var url = 'resources/worker-no-op.js';
    var scope = 'resources/blank.html';
    service_worker_unregister_and_register(t, url, scope).then(t.step_func(onRegister));

    function onRegister(worker) {
        worker.addEventListener('statechange', t.step_func(function(event) {
            if (event.target.state == 'activated')
                onActive();
        }));
    }

    function onActive() {
        with_iframe(scope, t.step_func(function(frame) {
            var w = frame.contentWindow;
            var controller = w.navigator.serviceWorker.controller;
            assert_true(controller instanceof w.ServiceWorker,
                        'controller should be a ServiceWorker object');
            assert_equals(controller.scriptURL, normalizeURL(url));
            service_worker_unregister_and_done(t, scope);
        }));
    }
});
</script>
</body>
