third_party/WebKit/LayoutTests/virtual/serviceworker/http/tests/serviceworker/controller-on-load-expected.txt


This is a testharness.js-based test.
PASS controller is set for a controlled document 
Harness: the test ran to completion.
